Old script from 2015-05 that reads in the itunes' library xml file and converts it into a json format (commented out).
The last iteration block goes and outputs the song and play count.  The commented out code was to create a list of 0-sized mp3 files that were missing using the location property.  This was to recover the list of songs that was in dead secondary hard drive that is not in the current collection (that paths were manipulated in the xml file to point at the current media location).